+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Enforce mesa software rendering even when nvidia driver is loaded
 * Catch when ffmpeg process exited
 * Add more hooks to prevent mouse capture
+* Implement fake window position instead of 0,0. Also update window size
 
 ### Changed